Add darwin_annotation_type to color look up and move Darwin util function to utils
